Used eclipse code cleanup on all source files.

+++ src/test/java/nom/tam/util/test/ArrayFuncs2Test.java
@@ -7,12 +7,12 @@ package nom.tam.util.test;
  * Copyright (C) 2004 - 2021 nom-tam-fits
  * %%
  * This is free and unencumbered software released into the public domain.
- * 
+ *
  * Anyone is free to copy, modify, publish, use, compile, sell, or
  * distribute this software, either in source code form or as a compiled
  * binary, for any purpose, commercial or non-commercial, and by any
  * means.
- * 
+ *
  * In jurisdictions that recognize copyright laws, the author or authors
  * of this software dedicate any and all copyright interest in the
  * software to the public domain. We make this dedication for the benefit
@@ -20,7 +20,7 @@ package nom.tam.util.test;
  * successors. We intend this dedication to be an overt act of
  * relinquishment in perpetuity of all present and future rights to this
  * software under copyright law.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
@@ -35,10 +35,10 @@ import static org.junit.Assert.assertEquals;
 
 import java.util.Arrays;
 
+import org.junit.Test;
+
 import nom.tam.util.ArrayFuncs;
 import nom.tam.util.TestArrayFuncs;
-
-import org.junit.Test;
 
 public class ArrayFuncs2Test {
 
@@ -127,9 +127,9 @@ public class ArrayFuncs2Test {
         assertEquals("flatten()", 1, dims.length);
 
         int[] newdims = {
-            8,
-            9,
-            10
+                8,
+                9,
+                10
         };
 
         int[][][] test7 = (int[][][]) ArrayFuncs.curl(test6, newdims);
@@ -155,10 +155,10 @@ public class ArrayFuncs2Test {
         assertEquals("newInstance(vector)", 1, xtd.length);
         assertEquals("newInstance(vector)", 20, xtd[0]);
         double[][][][] xtest4 = (double[][][][]) ArrayFuncs.newInstance(double.class, new int[]{
-            5,
-            4,
-            3,
-            2
+                5,
+                4,
+                3,
+                2
         });
         xtd = ArrayFuncs.getDimensions(xtest4);
         assertEquals("newInstance(tensor)", 4, xtd.length);
@@ -178,11 +178,11 @@ public class ArrayFuncs2Test {
         assertEquals("getBaseArray()", 2, xtest4x.length);
 
         double[] x = {
-            1,
-            2,
-            3,
-            4,
-            5
+                1,
+                2,
+                3,
+                4,
+                5
         };
         double[] y = new double[x.length];
         for (int i = 0; i < y.length; i += 1) {
@@ -196,11 +196,11 @@ public class ArrayFuncs2Test {
         assertEquals("eqTest5", false, TestArrayFuncs.arrayEquals(x, y, 1.E-5, 0.));
 
         float[] fx = {
-            1,
-            2,
-            3,
-            4,
-            5
+                1,
+                2,
+                3,
+                4,
+                5
         };
         float[] fy = new float[fx.length];
         for (int i = 0; i < fy.length; i += 1) {
@@ -303,7 +303,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -312,7 +312,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((long) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -321,7 +321,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -387,7 +387,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -396,7 +396,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((long) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -405,7 +405,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -471,7 +471,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -480,7 +480,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((long) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -489,7 +489,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -555,7 +555,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -573,7 +573,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -723,7 +723,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -732,7 +732,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((long) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -741,7 +741,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -807,7 +807,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -816,7 +816,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((long) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -825,7 +825,7 @@ public class ArrayFuncs2Test {
             ArrayFuncs.copyInto(deepArray, copy);
             for (int index1 = 0; index1 < copy.length; index1++) {
                 for (int index2 = 0; index2 < copy[index1].length; index2++) {
-                    assertEquals((float) deepArray[index1][index2], copy[index1][index2], 0.00000001);
+                    assertEquals(deepArray[index1][index2], copy[index1][index2], 0.00000001);
                 }
             }
         }
@@ -834,15 +834,15 @@ public class ArrayFuncs2Test {
     @Test
     public void testReverse() {
         int[] data = {
-            1,
-            2,
-            3,
-            4,
-            5,
-            6,
-            7,
-            8,
-            9
+                1,
+                2,
+                3,
+                4,
+                5,
+                6,
+                7,
+                8,
+                9
         };
         int[] dataReversed = ArrayFuncs.reverseIndices(data);
         assertEquals("[9, 8, 7, 6, 5, 4, 3, 2, 1]", Arrays.toString(dataReversed));
